+++ integration-tests/e2e-tests/src/abilities/WalletSdk.ts
@@ -1,5 +1,6 @@
 import { Ability, Discardable, Initialisable, Interaction, Question, QuestionAdapter } from "@serenity-js/core"
-import SDK from "@hyperledger/identus-sdk"
+import SDK from "@hyperledger/identus-sdk";
+import * as Anoncreds from "@hyperledger/identus-sdk/plugins/anoncreds";
 import axios from "axios"
 import { CloudAgentConfiguration } from "../configuration/CloudAgentConfiguration"
 import { randomUUID, UUID } from "crypto"
@@ -102,8 +103,8 @@ export class WalletSdk extends Ability implements Initialisable, Discardable {
       pluto,
       mediatorDID,
       castor
-    })
-    this.sdk.plugins.register(SDK.Plugins.Anoncreds)
+    });
+    this.sdk.plugins.register(Anoncreds.plugin);
 
     this.sdk.addListener(
       SDK.ListenerKey.MESSAGE, async (messages: SDK.Domain.Message[]) => {
