+++ src/plugins/internal/anoncreds/CredentialIssue.ts
@@ -4,7 +4,7 @@ import { AnonCredsCredential } from "../../../pollux/models/AnonCredsVerifiableC
 import { expect } from "../../../utils";
 import { GetLinkSecret } from "./GetLinkSecret";
 import { FetchCredentialDefinition } from "./FetchCredentialDefinition";
-import type { Context } from "./index";
+import type { Context } from "./plugin";
 import * as Types from "./types";
 import { Plugins } from "../../../plugins";
