+++ src/plugins/internal/anoncreds/CredentialOffer.ts
@@ -3,7 +3,7 @@ import { expect } from "../../../utils";
 import { Payload } from "../../../domain/protocols/Payload";
 import { GetLinkSecret } from "./GetLinkSecret";
 import { FetchCredentialDefinition } from "./FetchCredentialDefinition";
-import type { Context } from "./index";
+import type { Context } from "./plugin";
 import * as Types from "./types";
 import { Plugins } from "../../../plugins";
