+++ src/plugins/internal/anoncreds/FetchSchema.ts
@@ -1,5 +1,5 @@
-import { AgentContext } from "../../../edge-agent/didcomm/Context";
 import { Task } from "../../../utils";
+import type { Context } from "./plugin";
 import * as Types from "./types";
 
 interface Args {
@@ -7,7 +7,7 @@ interface Args {
 }
 
 export class fetchSchema extends Task<Types.Schema, Args> {
-  async run(ctx: AgentContext) {
+  async run(ctx: Context) {
     const response = await ctx.Api.request("GET", this.args.uri);
     // [ ] validate <Anoncreds.CredentialSchemaType>
     return response.body as Types.Schema;
