+++ tsup.config.js
@@ -3,6 +3,7 @@ import { defineConfig } from 'tsup'
 export default defineConfig({
   entry: [
     'src/index.ts',
+    'src/plugins/internal/anoncreds/index.ts',
     'src/plugins/internal/oidc/index.ts',
   ],
   outDir: "build",
